Fix triangleStart overflow in M2LegacyLoader _parse_views_inline
The JS code does: triangleStart += level << 16
The C++ had: sm.triangleStart += static_cast<uint16_t>(sm.level << 16)

static_cast<uint16_t>(level << 16) is always 0 because the lower 16 bits
of any value shifted left by 16 are all zeros. This made the entire
operation a no-op, losing the level information entirely.

Fix: widen triangleStart from uint16_t to uint32_t (matching JS number
semantics) and remove the erroneous truncating cast.

+++ src/js/3D/loaders/M2LegacyLoader.cpp
@@ -447,7 +447,7 @@ void M2LegacyLoader::_parse_views_inline(uint32_t ofs) {
 				sm.sortRadius = data.readFloatLE();
 			}
 
-			sm.triangleStart += static_cast<uint16_t>(sm.level << 16);
+			sm.triangleStart += sm.level << 16;
 		}
 
 		// read texture units
